refactor surface connections into their own file
do cleanup for on_surface_deleted
