+++ control.lua
@@ -19,6 +19,7 @@ require "script.stop-update"
 require "script.dispatcher"
 require "script.stop-events"
 require "script.train-events"
+require "script.surface-connections"
 require "script.interface" -- ties into other modules
 require "script.init" -- requires other modules loaded first
