+++ script/interface.lua
@@ -17,47 +17,6 @@ on_provider_unscheduled_cargo_alert = script.generate_event_name()
 on_requester_unscheduled_cargo_alert = script.generate_event_name()
 on_requester_remaining_cargo_alert = script.generate_event_name()
 
-local function clear_all_surface_connections()
-  global.ConnectedSurfaces = {}
-end
-
-local function sorted_pair(number1, number2)
-  return (number1 < number2) and (number1..'|'..number2) or (number2..'|'..number1)
-end
-
-local function lazy_subtable(a_table, key)
-  local subtable = a_table[key]
-  if not subtable then
-    subtable = {}
-    a_table[key] = subtable
-  end
-  return subtable
-end
-
-local function disconnect_surfaces(entity1, entity2)
-  local surface_pair_key = sorted_pair(entity1.surface.index, entity2.surface.index)
-  local surface_connections = global.ConnectedSurfaces[surface_pair_key]
-
-  if surface_connections then
-    surface_connections[sorted_pair(entity1.unit_number, entity2.unit_number)] = nil
-  end
-end
-
-local function connect_surfaces(entity1, entity2, network_id)
-  if entity1.surface == entity2.surface then
-    error("the given entities must not be on the same surface")
-  end
-
-  local surface_pair_key = sorted_pair(entity1.surface.index, entity2.surface.index)
-  local surface_connections = lazy_subtable(global.ConnectedSurfaces, surface_pair_key)
-
-  local entity_pair_key = sorted_pair(entity1.unit_number, entity2.unit_number)
-  surface_connections[entity_pair_key] = {
-    entity1 = entity1,
-    entity2 = entity2,
-    network_id = network_id,
-  }
-end
 
 
 
@@ -89,36 +48,39 @@ remote.add_interface("logistic-train-network", {
   -- The connection has a network_id that acts as an additional mask that is applied to potential providers,
   -- so using -1 here would apply no further restrictions and using 0 would make the connection unusable.
   -- The connection is bi-directional but not transitive, i.e. surface A -> B implies B -> A, but A -> B and B -> C does not imply A -> C.
+  --
   -- LTN has no notion of how a train actually moves between surfaces.
   -- It is the caller's responsibility to modify the train schedule accordingly, most appropriately in an on_delivery_created_event.
   -- Its field 'surface_connections' will be the table of connections that had a matching network_id and force.
   -- For intra-surface deliveries that field will be an empty table.
+  --
   -- Trains will be sourced from the provider surface only and should return to it after leaving the requester.
   -- This is to avoid creating an imbalance in the train pools per surface.
-  connect_surfaces = connect_surfaces, -- function(entity1 :: LuaEntity, entity2 :: LuaEntity)
+  connect_surfaces = ConnectSurfaces, -- function(entity1 :: LuaEntity, entity2 :: LuaEntity, network_id :: uint)
 
   -- Explicitly severs the surface connection that is formed by the two given entities.
   -- Active deliveries will not be affected.
   -- Calling this function for two entities that are not connected has no effect.
   -- It is not necessary to call this function if one or both of the entities are deleted.
-  disconnect_surfaces = disconnect_surfaces, -- function(entity1 :: LuaEntity, entity2 :: LuaEntity)
+  disconnect_surfaces = DisconnectSurfaces, -- function(entity1 :: LuaEntity, entity2 :: LuaEntity)
 
   -- Clears all surface connections.
   -- Active deliveries will not be affected
-  -- This function exists mostly for debugging purposes, so there is no event that is raised afterwards.
-  clear_all_surface_connections = clear_all_surface_connections,
-
-  -- TODO handle on_surface_deleted
+  -- This function exists mostly for debugging purposes, there is no event that is raised to tell other mods about this.
+  -- TODO remove this?
+  clear_all_surface_connections = ClearAllSurfaceConnections,
 
   -- Re-assigns a delivery to a different train.
   -- Scripts should call this function if they create a train based on another train, for example when moving a train to a different surface.
   -- It is the caller's responsibility to make sure that the new train's schedule matches the old one's before calling this function.
+  -- Otherwise LTN won't be able to add missing temporary stops for logistic stops that are now on the same surface as the train.
   -- Calling this function with an old_train_id that is not executing a delivery has no effect.
   -- It is not necessary to call this function when coupling trains via script, LTN is already aware of that through Factorio events.
   reassign_delivery = ReassignDelivery,
 
-  -- Allow to keep surfaces mostly separate with a few explicit deliveries between them without the need to set the network_id explicitly on each stop on both surfaces.
-  -- TODO set_surface_default_network_id(surface :: LuaSurface, uint network_id)
+  -- TODO Is this worth it? It's a stretch goal and a completely separate feature in either case
+  -- Allows to keep surfaces mostly separate with a few explicit deliveries between them without the need to set the network_id explicitly on each stop on both surfaces.
+  -- set_surface_default_network_id(surface :: LuaSurface, uint network_id)
 })
